refactor: rename controllers and routes and remove new prefixes
