--- src/controllers/newUserController.ts
@@ -1,115 +1,0 @@
-import { NextFunction, Request, Response } from "express"
-import {
-  deleteUserService,
-  getUserService,
-  getUsersService,
-  updateUserService,
-} from "../services/newUserService"
-import { NewRole } from "../utils/types"
-import {
-  getUsersSchema,
-  ProfileSchema,
-  updateUserSchema,
-} from "../validations/newUserValidation"
-
-export const getProfile = async (
-  req: any,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const userId = req.user.id
-    const user = await getUserService({ _id: userId })
-    return res.status(200).send(user)
-  } catch (error) {
-    return next(error)
-  }
-}
-
-export const getUsersController = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const search = req.query
-    await getUsersSchema.validateAsync(search)
-    const users = await getUsersService(search)
-    return res.status(200).send(users)
-  } catch (error) {
-    return next(error)
-  }
-}
-
-export const updateProfile = async (
-  req: any,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const userId = req.user.id
-    await ProfileSchema.validateAsync(req.body)
-
-    const user = await updateUserService(userId, req.body)
-    return res.status(200).send(user)
-  } catch (error) {
-    return next(error)
-  }
-}
-
-export const updateUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const userId = req.params.userId
-    const { name, role } = req.body
-    await updateUserSchema.validateAsync({ name, role })
-
-    const user = await updateUserService(userId, { name, role })
-    return res.status(200).send(user)
-  } catch (error) {
-    return next(error)
-  }
-}
-
-export const deleteUser = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const userId = req.params.userId
-    await deleteUserService(userId)
-    return res.status(200).send("User deleted successfully")
-  } catch (error) {
-    return next(error)
-  }
-}
-
-export const toggleUserActiveStatus = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-) => {
-  try {
-    const userId = req.params.userId
-
-    // Get the user to check if they're admin or coach
-    const targetUser = await getUserService({ _id: userId })
-    if (![NewRole.Admin, NewRole.Coach].includes(targetUser.role)) {
-      return res.status(403).json({
-        message: "You can only activate/deactivate admin and coach users",
-      })
-    }
-
-    const updatedUser = await updateUserService(userId, {
-      active: !targetUser.active,
-    })
-
-    return res.status(200).send(updatedUser)
-  } catch (error) {
-    return next(error)
-  }
-}
