+++ src/controllers/responseController.ts
@@ -3,10 +3,6 @@ import {
   createApplicantResponseValidation,
   createCoachResponseValidation,
 } from "../validations/responseValidation"
-import {
-  createApplicantResponseService,
-  createCoachResponseService,
-} from "../services/responseService"
 
 export const createCoachResponse = async (
   req: any,
@@ -14,13 +10,7 @@ export const createCoachResponse = async (
   next: NextFunction,
 ) => {
   try {
-    const loggedInUser = req.user
-    await createCoachResponseValidation.validateAsync(req.body)
-    const createdResponse = await createCoachResponseService(
-      loggedInUser,
-      req.body,
-    )
-    return res.status(201).json(createdResponse)
+    return await createCoachResponseValidation.validateAsync(req.body)
   } catch (error) {
     return next(error)
   }
@@ -32,14 +22,9 @@ export const createApplicantResponse = async (
   next: NextFunction,
 ) => {
   try {
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const loggedInUser = req.user
-    await createApplicantResponseValidation.validateAsync(req.body)
-    const createdResponse = await createApplicantResponseService(
-      loggedInUser,
-      req.body,
-      req.query.action === "submit",
-    )
-    return res.status(201).json(createdResponse)
+    return await createApplicantResponseValidation.validateAsync(req.body)
   } catch (error) {
     return next(error)
   }
