--- src/middlewares/newAuthenticate.ts
@@ -1,44 +1,0 @@
-import { NextFunction, Response } from "express"
-import { verify } from "jsonwebtoken"
-import dotenv from "dotenv"
-import { getUserService } from "../services/newUserService"
-
-dotenv.config()
-
-const accessKey = process.env.ACCESS_TOKEN_KEY || ""
-
-export const verifyJWT = (req: any, res: Response, next: NextFunction) => {
-  const authHeader = req.headers.authorization
-  if (!authHeader?.startsWith("Bearer ")) {
-    return res.status(401).send("User not logged in")
-  }
-
-  const token = authHeader.split(" ")[1]
-
-  return verify(token, accessKey, async (err: any, decoded: any) => {
-    if (err) {
-      return res.status(403).json({
-        type: "JsonWebTokenError",
-        errorMessage: "Invalid access token",
-      })
-    }
-
-    try {
-      const user = await getUserService({ _id: decoded.id })
-
-      // Check if user is active
-      if (!user.active) {
-        return res.status(403).json({
-          type: "DeactivatedAccount",
-          errorMessage:
-            "Account was deactivated, please consult the admin for more information",
-        })
-      }
-
-      req.user = user
-      return next()
-    } catch (err) {
-      return next(err)
-    }
-  })
-}
