+++ src/server.ts
@@ -10,19 +10,15 @@ import swaggerUI from "swagger-ui-express"
 import YAML from "yamljs"
 import CustomError from "./middlewares/customError"
 import { errorHandler } from "./middlewares/errorHandler"
-import newAuthRoute from "./routes/newAuthRoute"
 import authRoute from "./routes/authRoute"
 import coachRoute from "./routes/coachRoute"
-import cohortRoutes from "./routes/cohortRoutes"
-import newCohortRoutes from "./routes/newCohortRoute"
+import cohortRoutes from "./routes/cohortRoute"
 import formRoute from "./routes/formRoute"
 import joinWaitListRoute from "./routes/joinWaitListRoute"
 import overviewRoute from "./routes/overviewRoute"
-import participantRoutes from "./routes/participantRoutes"
 import questionRoute from "./routes/questionRoute"
 import responseRoute from "./routes/responseRoute"
 import traineeRoute from "./routes/traineeRoute"
-import newUserRoute from "./routes/newUserRoute"
 import userRoute from "./routes/userRoute"
 import { URL_NOT_FOUND } from "./utils/errorCodes"
 import { htmlDocumentationResponse } from "./utils/helpers/htmlDocumentationResponse"
@@ -86,15 +82,11 @@ app.use("/api-docs", swaggerUI.serve, swaggerUI.setup(swaggerDocumentation))
 
 app.use(cors({ origin: "*", credentials: true }))
 app.use(express.json())
-app.use("/new-auth", newAuthRoute)
 app.use("/auth", authRoute)
 app.use("/users", userRoute)
-app.use("/new-users", newUserRoute)
 app.use("/trainees", traineeRoute)
-app.use("/participants", participantRoutes)
 app.use("/coaches", coachRoute)
 app.use("/cohorts", cohortRoutes)
-app.use("/new-cohorts", newCohortRoutes)
 app.use("/forms", formRoute)
 app.use("/questions", questionRoute)
 app.use("/responses", responseRoute)
