+++ src/services/traineeService.ts
@@ -1,13 +1,13 @@
 import CustomError from "../middlewares/customError"
 import User from "../models/User"
-import { USER_NOT_FOUND } from "../utils/errorCodes"
+import Trainee from "../models/Trainee"
 import {
   getTraineesForCoachQuery,
   getTraineesQuery,
 } from "../queries/traineesQuery"
-import { getCohortService } from "./cohortService"
-import { updateUserService } from "./userService"
+import { USER_NOT_FOUND } from "../utils/errorCodes"
 import { updateUserDto } from "../utils/types"
+import { updateUserService } from "./userService"
 
 export const getTraineesService = async ({
   searchString,
@@ -44,31 +44,15 @@ export const updateTraineeService = async (
   traineeId: string,
   updates: updateUserDto,
 ) => {
-  const currentCohort = await getCohortService({ isActive: true })
-  const applicant = currentCohort.applicants.find(
-    (applicant) => applicant.id.toString() === traineeId,
-  )
+  // To be worked on when new trainee services are implemented
+  const trainee = await Trainee.findById(traineeId)
 
-  if (!applicant) {
+  if (!trainee) {
     throw new CustomError(
       USER_NOT_FOUND,
-      "Applicant not found in the current cohort",
+      "Trainee not found in the current cohort",
       404,
     )
-  }
-
-  if (updates.coach) {
-    const coach = currentCohort.coaches.find(
-      (coach) => coach.toString() === coach,
-    )
-
-    if (!coach) {
-      throw new CustomError(
-        USER_NOT_FOUND,
-        "Coach not found in the current cohort",
-        404,
-      )
-    }
   }
 
   return updateUserService(traineeId, updates)
